move chars_cmp and chars_next_cmp to its own modules
